Updated the UI to hide diagnosis and sick leave's edit and delete buttons for doctors that don't belong to the patient's appointment and remained available for doctors that are part of the appointment model as well as for admins.

+++ lib/widgets/diagnosis_dialog.dart
@@ -3,6 +3,7 @@ import 'package:intl/intl.dart';
 import 'package:provider/provider.dart';
 import '../domain/diagnosis.dart';
 import '../provider/appointment_provider.dart';
+import '../provider/auth_provider.dart';
 import 'edit_diagnosis_form.dart';
 import 'role_based_widget.dart';
 import 'treatment_dialog.dart';
@@ -10,8 +11,14 @@ import 'treatment_dialog.dart';
 class DiagnosisDialog extends StatefulWidget {
   final List<Diagnosis> diagnoses;
   final int appointmentId;
+  final String doctorKeycloakUserId;
 
-  const DiagnosisDialog({Key? key, required this.diagnoses, required this.appointmentId}) : super(key: key);
+  const DiagnosisDialog({
+    Key? key,
+    required this.diagnoses,
+    required this.appointmentId,
+    required this.doctorKeycloakUserId,
+  }) : super(key: key);
 
   @override
   _DiagnosisDialogState createState() => _DiagnosisDialogState();
@@ -46,6 +53,9 @@ class _DiagnosisDialogState extends State<DiagnosisDialog> {
 
   @override
   Widget build(BuildContext context) {
+    final authProvider = Provider.of<AuthProvider>(context, listen: false);
+    final currentUserKeycloakId = authProvider.keycloakUserId;
+
     return AlertDialog(
       title: const Text('Diagnosis Details'),
       content: SizedBox(
@@ -65,38 +75,40 @@ class _DiagnosisDialogState extends State<DiagnosisDialog> {
                 child: Row(
                   mainAxisSize: MainAxisSize.min,
                   children: [
-                    ElevatedButton(
-                      onPressed: () {
-                        showDialog(
-                          context: context,
-                          builder: (_) => EditDiagnosisForm(
-                            appointmentId: widget.appointmentId,
-                            diagnosis: diagnosis,
-                            onUpdate: _updateDiagnosis,
-                          ),
-                        );
-                      },
-                      child: const Text('Edit'),
-                    ),
-                    const SizedBox(width: 8),
-                    ElevatedButton(
-                      onPressed: () async {
-                        bool success = await Provider.of<AppointmentProvider>(context, listen: false)
-                            .deleteDiagnosis(widget.appointmentId, diagnosis.id);
-                        if (success) {
-                          _onDeleteDiagnosis(diagnosis.id);
-                        } else {
-                          ScaffoldMessenger.of(context).showSnackBar(
-                            const SnackBar(content: Text('Failed to delete diagnosis')),
+                    if (authProvider.roles.contains('admin') || currentUserKeycloakId == widget.doctorKeycloakUserId)
+                      ElevatedButton(
+                        onPressed: () {
+                          showDialog(
+                            context: context,
+                            builder: (_) => EditDiagnosisForm(
+                              appointmentId: widget.appointmentId,
+                              diagnosis: diagnosis,
+                              onUpdate: _updateDiagnosis,
+                            ),
                           );
-                        }
-                      },
-                      style: ElevatedButton.styleFrom(
-                        backgroundColor: Colors.red,
-                        foregroundColor: Colors.white,
+                        },
+                        child: const Text('Edit'),
                       ),
-                      child: const Text('Delete'),
-                    ),
+                    const SizedBox(width: 8),
+                    if (authProvider.roles.contains('admin') || currentUserKeycloakId == widget.doctorKeycloakUserId)
+                      ElevatedButton(
+                        onPressed: () async {
+                          bool success = await Provider.of<AppointmentProvider>(context, listen: false)
+                              .deleteDiagnosis(widget.appointmentId, diagnosis.id);
+                          if (success) {
+                            _onDeleteDiagnosis(diagnosis.id);
+                          } else {
+                            ScaffoldMessenger.of(context).showSnackBar(
+                              const SnackBar(content: Text('Failed to delete diagnosis')),
+                            );
+                          }
+                        },
+                        style: ElevatedButton.styleFrom(
+                          backgroundColor: Colors.red,
+                          foregroundColor: Colors.white,
+                        ),
+                        child: const Text('Delete'),
+                      ),
                   ],
                 ),
               ),
